requirements.txt
click==8.0.3
Flask==2.2.2
gunicorn==20.1.0
importlib-metadata==6.0.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.1.2
python-dotenv==0.21.1
Werkzeug==2.2.2
zipp==3.11.0
flask_wtf
flask_mail
email_validator